requirements/rag2_requirements.txt
altair==5.5.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
cachetools==6.2.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046050867/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1754767332901/work
click==8.3.0
cmake==4.1.0
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1751548120229/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography-split_1701563205069/work
diskcache==5.6.3
exceptiongroup==1.3.0
faiss-gpu==1.7.2
filelock==3.13.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hf-xet==1.1.10
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
huggingface-hub==0.23.2
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
libmambapy @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1749108038998/work/libmambapy
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
mpmath==1.3.0
matplotlib==3.10.7
narwhals==2.10.2
networkx==3.3
ninja==1.13.0
nltk==3.9.2
numpy < 2
#nvidia-cublas-cu12==12.8.4.1
#nvidia-cuda-cupti-cu12==12.8.90
#nvidia-cuda-nvrtc-cu12==12.8.93
#nvidia-cuda-runtime-cu12==12.8.90
#nvidia-cudnn-cu12==9.10.2.21
#nvidia-cufft-cu12==11.3.3.83
#nvidia-cufile-cu12==1.13.1.3
#nvidia-curand-cu12==10.3.9.90
#nvidia-cusolver-cu12==11.7.3.90
#nvidia-cusparse-cu12==12.5.8.93
#nvidia-cusparselt-cu12==0.7.1
#nvidia-nccl-cu12==2.27.3
#nvidia-nvjitlink-cu12==12.8.93
#nvidia-nvtx-cu12==12.8.90
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.3
pathspec==0.12.1
pillow==10.3.0
protobuf==6.33.0
pyarrow==21.0.0
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1649384799612/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydeck==0.9.1
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1685514481738/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.9.18
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
rpds-py==0.28.0
ruamel-yaml-conda @ file:///home/conda/feedstock_root/build_artifacts/ruamel_yaml_1636009152712/work
safetensors==0.4.3
scikit-learn==1.7.2
scikit_build_core==0.11.6
scipy==1.15.3
sentence-transformers==2.6.1
sentencepiece==0.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.51.0
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
torch==2.1.2+cu121
torchaudio==0.11.0+cu113
torchvision==0.16.2+cu121
tornado==6.5.2
tqdm==4.67.1
transformers==4.43.3
triton==2.1.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
uvicorn==0.37.0
watchdog==6.0.0
zstandard @ file:///croot/zstandard_1677013143055/work
